+++ hs_access_control/tests/test_undo.py
@@ -1,126 +1,18 @@
-__author__ = 'Alva'
-
-import unittest
-from django.http import Http404
 from django.test import TestCase
-from django.utils import timezone
-from django.core.exceptions import PermissionDenied
-from django.contrib.auth.models import User, Group
-from pprint import pprint
+from django.contrib.auth.models import Group
 
-from hs_access_control.models import UserAccess, GroupAccess, ResourceAccess, \
-    UserResourceProvenance, UserResourcePrivilege, \
+from hs_access_control.models import UserResourceProvenance, UserResourcePrivilege, \
     GroupResourceProvenance, GroupResourcePrivilege, \
     UserGroupProvenance, UserGroupPrivilege, \
     PrivilegeCodes, ProvenanceCodes
 
 from hs_core import hydroshare
-from hs_core.models import GenericResource, BaseResource
 from hs_core.testing import MockIRODSTestCaseMixin
 
-from hs_access_control.tests.utilities import *
+from hs_access_control.tests.utilities import is_equal_to_as_set, global_reset, \
+    check_provenance_synchronization
 
-
-def check_provenance_synchronization(self):
-    for u in User.objects.all():
-        for r in BaseResource.objects.all():
-            prov = UserResourceProvenance.get_privilege(resource=r, user=u)
-            priv = UserResourcePrivilege.get_privilege(resource=r, user=u)
-            self.assertEqual(prov, priv, 
-                             str.format("prov={}, priv={}, resource={}, user={}", prov, priv, r, u))
-    for u in User.objects.all():
-        for g in Group.objects.all():
-            prov = UserGroupProvenance.get_privilege(group=g, user=u)
-            priv = UserGroupPrivilege.get_privilege(group=g, user=u)
-            self.assertEqual(prov, priv, 
-                             str.format("prov={}, priv={}, group={}, user={}", prov, priv, g, u))
-    for g in Group.objects.all():
-        for r in BaseResource.objects.all():
-            prov = GroupResourceProvenance.get_privilege(resource=r, group=g)
-            priv = GroupResourcePrivilege.get_privilege(resource=r, group=g)
-            self.assertEqual(prov, priv, 
-                             str.format("prov={}, priv={}, group={}, resource={}", prov, priv, g, r))
-
-
-def printGroupResourceState():
-    print "==================================="
-    print "GroupResourcePrivilege"
-    priv = GroupResourcePrivilege.objects.all().order_by('group__id', 'resource__id')
-    o = None
-    for p in priv:
-        if o is not None and (p.group != o.group or p.resource != o.resource):
-            print "------------------------------"
-        print(p)
-        o = p
-    print "==================================="
-    print "GroupResourceProvenance"
-    prov = GroupResourceProvenance.objects.all().order_by(
-        'group__id', 'resource__id', 'start')
-    o = None
-    for p in prov:
-        if o is not None and (p.group != o.group or p.resource != o.resource):
-            print "------------------------------"
-        current = GroupResourceProvenance.get_current_record(
-            resource=p.resource, group=p.group)
-        star = ''
-        if current == p:
-            star = 'CURRENT'
-        print(p, star)
-        o = p
-
-
-def printUserResourceState():
-    print "==================================="
-    print "UserResourcePrivilege"
-    priv = UserResourcePrivilege.objects.all().order_by('user__id', 'resource__id')
-    o = None
-    for p in priv:
-        if o is not None and (p.user != o.user or p.resource != o.resource):
-            print "------------------------------"
-        print(p)
-        o = p
-    print "==================================="
-    print "UserResourceProvenance"
-    prov = UserResourceProvenance.objects.all().order_by(
-        'user__id', 'resource__id', 'start')
-    o = None
-    for p in prov:
-        if o is not None and (p.user != o.user or p.resource != o.resource):
-            print "------------------------------"
-        current = UserResourceProvenance.get_current_record(
-            resource=p.resource, user=p.user)
-        star = ''
-        if current == p:
-            star = 'CURRENT'
-        print(p, star)
-        o = p
-
-
-def printUserGroupState():
-    print "==================================="
-    print "UserGroupPrivilege"
-    priv = UserGroupPrivilege.objects.all().order_by('user__id', 'group__id')
-    o = None
-    for p in priv:
-        if o is not None and (p.user != o.user or p.group != o.group):
-            print "------------------------------"
-        pprint(p)
-        o = p
-    print "==================================="
-    print "UserGroupProvenance"
-    prov = UserGroupProvenance.objects.all().order_by(
-        'user__id', 'group__id', 'start')
-    o = None
-    for p in prov:
-        if o is not None and (p.user != o.user or p.group != o.group):
-            print "------------------------------"
-        current = UserGroupProvenance.get_current_record(
-            group=p.group, user=p.user)
-        star = ''
-        if current == p:
-            star = 'CURRENT'
-        print(p, star)
-        o = p
+__author__ = 'Alva'
 
 
 class UnitTests(MockIRODSTestCaseMixin, TestCase):
@@ -223,7 +115,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # George grants Alva privilege
@@ -253,7 +145,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # Alva grants John privilege
@@ -289,15 +181,14 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.user, john)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, alva)
 
         # now George overrides Alva on John's privilege
         george.uaccess.share(
             resource=bikes,
             user=john,
-            privilege=PrivilegeCodes.VIEW,
-            grantor=george)
+            privilege=PrivilegeCodes.VIEW)
 
         self.assertEqual(
             UserResourceProvenance.get_privilege(
@@ -325,7 +216,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.user, john)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # Crosstalk test: George grants Alva privilege over harps
@@ -355,7 +246,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # check new privileges: should be independent.
@@ -380,12 +271,8 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, harps)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
-
-        prov = UserResourceProvenance.objects.all().order_by('start')
-        for p in prov:
-            pprint(p)
 
         check_provenance_synchronization(self)
 
@@ -424,7 +311,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # George grants Alva privilege
@@ -454,7 +341,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # Alva grants John privilege
@@ -490,7 +377,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.user, john)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, alva)
 
         # now George overrides Alva on John's privilege
@@ -525,7 +412,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.user, john)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # Crosstalk test: George grants Alva privilege over harpers
@@ -555,7 +442,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # check new privileges: should be independent of old privileges
@@ -580,7 +467,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, harpers)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         check_provenance_synchronization(self)
@@ -621,7 +508,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # George grants Harpers change privilege
@@ -650,10 +537,22 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.group, harpers)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
-        # Alva downgrades Harpers privilege
+        # Alva is a harper
+        george.uaccess.share(
+            group=harpers,
+            user=alva,
+            privilege=PrivilegeCodes.CHANGE)
+
+        # Alva can access bikes
+        george.uaccess.share(
+            resource=bikes,
+            user=alva,
+            privilege=PrivilegeCodes.CHANGE)
+
+        # Alva downgrades Harpers privilege on bikes
         alva.uaccess.share(
             resource=bikes,
             group=harpers,
@@ -686,30 +585,30 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.group, harpers)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, alva)
 
-        # now George overrides Alva on  Harpers privilege
+        # now George overrides Alva on Harpers privilege
         george.uaccess.share(
             resource=bikes,
             group=harpers,
-            privilege=PrivilegeCodes.VIEW)
+            privilege=PrivilegeCodes.CHANGE)
 
         self.assertEqual(
             GroupResourceProvenance.get_privilege(
                 resource=bikes,
                 group=harpers),
-            PrivilegeCodes.VIEW)
+            PrivilegeCodes.CHANGE)
         self.assertEqual(
             GroupResourcePrivilege.get_privilege(
                 resource=bikes,
                 group=harpers),
-            PrivilegeCodes.VIEW)
+            PrivilegeCodes.CHANGE)
         self.assertTrue(
             is_equal_to_as_set(
                 GroupResourcePrivilege.get_undo_groups(
-                    resource=bikes, grantor=george), [
-                    bikers, harpers]))
+                    resource=bikes, grantor=george),
+                [bikers, harpers]))
         self.assertTrue(
             is_equal_to_as_set(
                 GroupResourcePrivilege.get_undo_groups(
@@ -720,8 +619,8 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
             resource=bikes, group=harpers)
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.group, harpers)
-        self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # Crosstalk test: George grants bikers privilege over harps
@@ -744,14 +643,14 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertTrue(
             is_equal_to_as_set(
                 GroupResourcePrivilege.get_undo_groups(
-                    resource=bikes, grantor=george), 
+                    resource=bikes, grantor=george),
                 [bikers, harpers]))
         record = GroupResourceProvenance.get_current_record(
             resource=bikes, group=bikers)
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # check new privileges: should be independent.
@@ -776,7 +675,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, harps)
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         check_provenance_synchronization(self)
@@ -837,7 +736,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # Roll back alva's privilege
@@ -865,7 +764,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.NONE)
-        self.assertEqual(record.state, ProvenanceCodes.INITIAL) 
+        self.assertEqual(record.state, ProvenanceCodes.INITIAL)
         self.assertEqual(record.grantor, None)
 
         # George grants Alva privilege
@@ -895,7 +794,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # Alva grants John privilege
@@ -931,7 +830,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.user, john)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, alva)
 
         # now George overrides Alva on John's privilege
@@ -966,15 +865,11 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.user, john)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # George changes mind and rolls back change
-        print("BEFORE UNDO(bikes, john, george)") 
-        printUserResourceState() 
         george.uaccess.undo_share(resource=bikes, user=john)
-        print("AFTER UNDO(bikes, john, george)") 
-        printUserResourceState() 
 
         # privilege has been rolled back
         self.assertEqual(
@@ -1004,7 +899,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.user, john)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.RESTORED) 
+        self.assertEqual(record.state, ProvenanceCodes.RESTORED)
         self.assertEqual(record.grantor, alva)
 
         # Crosstalk test: George grants Alva privilege over harps
@@ -1035,7 +930,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, bikes)
         self.assertEqual(record.user, john)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.RESTORED) 
+        self.assertEqual(record.state, ProvenanceCodes.RESTORED)
         self.assertEqual(record.grantor, alva)
 
         # check new privileges: should be independent.
@@ -1060,7 +955,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, harps)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # now roll back privilege over harps
@@ -1087,7 +982,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.resource, harps)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.NONE)
-        self.assertEqual(record.state, ProvenanceCodes.INITIAL) 
+        self.assertEqual(record.state, ProvenanceCodes.INITIAL)
         self.assertEqual(record.grantor, None)
 
         check_provenance_synchronization(self)
@@ -1148,11 +1043,11 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # Roll back alva's privilege
-        george.uaccess.undo_share(group=bikers, user=alva, grantor=george)
+        george.uaccess.undo_share(group=bikers, user=alva)
 
         self.assertEqual(
             UserGroupProvenance.get_privilege(
@@ -1176,7 +1071,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.NONE)
-        self.assertEqual(record.state, ProvenanceCodes.INITIAL) 
+        self.assertEqual(record.state, ProvenanceCodes.INITIAL)
         self.assertEqual(record.grantor, None)
 
         # George grants Alva privilege
@@ -1206,7 +1101,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # Alva grants John privilege
@@ -1242,7 +1137,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.user, john)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, alva)
 
         # now George overrides Alva on John's privilege
@@ -1277,7 +1172,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.user, john)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # George changes mind and rolls back change
@@ -1311,7 +1206,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.user, john)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.RESTORED) 
+        self.assertEqual(record.state, ProvenanceCodes.RESTORED)
         self.assertEqual(record.grantor, alva)
 
         # Crosstalk test: George grants Alva privilege over harpers
@@ -1342,7 +1237,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, bikers)
         self.assertEqual(record.user, john)
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
-        self.assertEqual(record.state, ProvenanceCodes.RESTORED) 
+        self.assertEqual(record.state, ProvenanceCodes.RESTORED)
         self.assertEqual(record.grantor, alva)
 
         # check new privileges: should be independent.
@@ -1367,7 +1262,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, harpers)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.VIEW)
-        self.assertEqual(record.state, ProvenanceCodes.REQUESTED) 
+        self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
 
         # now roll back privilege over harpers
@@ -1394,7 +1289,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.group, harpers)
         self.assertEqual(record.user, alva)
         self.assertEqual(record.privilege, PrivilegeCodes.NONE)
-        self.assertEqual(record.state, ProvenanceCodes.INITIAL) 
+        self.assertEqual(record.state, ProvenanceCodes.INITIAL)
         self.assertEqual(record.grantor, None)
 
         check_provenance_synchronization(self)
@@ -1514,6 +1409,18 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.privilege, PrivilegeCodes.CHANGE)
         self.assertEqual(record.state, ProvenanceCodes.REQUESTED)
         self.assertEqual(record.grantor, george)
+
+        # George grants Alva privilege
+        george.uaccess.share(
+            resource=bikes,
+            user=alva,
+            privilege=PrivilegeCodes.CHANGE)
+
+        # Alva is a harper
+        george.uaccess.share(
+            group=harpers,
+            user=alva,
+            privilege=PrivilegeCodes.CHANGE)
 
         # Alva grants harpers privilege
         alva.uaccess.share(
@@ -1677,7 +1584,7 @@ class UnitTests(MockIRODSTestCaseMixin, TestCase):
         self.assertEqual(record.grantor, george)
 
         # now roll back privilege over harps
-        george.uaccess.undo_share(resource=harps, group=bikers, grantor=george)
+        george.uaccess.undo_share(resource=harps, group=bikers)
 
         self.assertEqual(
             GroupResourceProvenance.get_privilege(
